# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fixly API
  version: 1.0.0
  description: |
    Public + admin HTTP API for Fixly. All responses follow the project's
    envelope contract: `{ "data": ... }` on success and
    `{ "error": { "code": string, "message": string } }` on failure. The
    socket.io surface (`booking:status`, `location:update`) is documented
    under the `Realtime` tag.
servers:
  - url: /api/v1
    description: Same-origin (proxied via Vite in dev, ingress in prod)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: phone must be a valid mobile number }
    Meta:
      type: object
      properties:
        total: { type: integer, example: 42 }
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
...
    Service:
      type: object
      properties:
        id: { type: string, example: elec }
        nameAr: { type: string, example: كهرباء }
        nameEn: { type: string, example: Electricity }
        descriptionAr: { type: string, nullable: true }
        priceJod: { type: string, example: '50.00' }
        durationMin: { type: integer, example: 45 }
        isActive: { type: boolean, example: true }
        category: { type: string, nullable: true, example: HOME_REPAIR }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [AWAITING_PAYMENT, PENDING, CONFIRMED, EN_ROUTE, ARRIVED, IN_PROGRESS, COMPLETED, CANCELLED, DISPUTED, NO_SHOW]
        scheduledAt: { type: string, format: date-time, nullable: true }
        totalJod: { type: string, example: '50.00' }
        service: { $ref: '#/components/schemas/Service' }
    CheckoutSession:
      type: object
      description: >
        Hosted-checkout session for a real PSP (HyperPay COPYandPAY). The client loads
        `scriptUrl?checkoutId=…` and renders a widget for the allowed card `brands`.
      properties:
        checkoutId: { type: string }
        scriptUrl: { type: string, format: uri }
        brands:
          type: array
          items: { type: string }
          example: [VISA, MASTER]
    CreateBookingInput:
      type: object
      required: [serviceId, addressLine, addressLat, addressLng]
      properties:
        serviceId: { type: string }
        addressLine: { type: string, minLength: 3, maxLength: 250 }
        addressLat: { type: number, minimum: -90, maximum: 90 }
        addressLng: { type: number, minimum: -180, maximum: 180 }
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: |
            ISO-8601 timestamp. Omit / set null for "now" (within 30 min).
...
    Technician:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        phone: { type: string }
        isVerified: { type: boolean }
        rating: { type: number, format: float, nullable: true }
        completedJobs: { type: integer }
...
    Subscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        planSlug: { type: string, example: protect }
        status: { type: string, enum: [ACTIVE, PAST_DUE, CANCELLED, EXPIRED] }
        priceJod: { type: string, example: '5.000' }
        discountPercent: { type: integer, example: 15 }
        guaranteeDays: { type: integer, example: 90 }
        priorityDispatch: { type: boolean }
        inspectionEveryDays: { type: integer, example: 90 }
        nextInspectionAt: { type: string, format: date-time, nullable: true }
        currentPeriodEnd: { type: string, format: date-time }
        paymentToken: { type: string, nullable: true, description: PSP card-on-file token — never a PAN. }
        cancelledAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
...
    BookingQuote:
      type: object
      description: >-
        Video pre-check (fixed_scope convenience) OR itemized quote-first
        (mandatory for quote_first services, §17.5) — `lines` is empty for the
        former, populated for the latter; labourFils/materialsFils are summed
        from `lines` for the itemized path.
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        serviceId: { type: string }
        status: { type: string, enum: [PENDING, QUOTED, ACCEPTED, DECLINED, EXPIRED] }
        videoUrl: { type: string, format: uri, nullable: true }
        siteMediaUrls: { type: array, items: { type: string, format: uri } }
        dimensionsNote: { type: string, nullable: true }
        requestedTier: { type: string, nullable: true, enum: [ECONOMY, STANDARD, PREMIUM] }
        description: { type: string, nullable: true }
        quotedJod: { type: string, nullable: true, example: '105.000' }
        labourFils: { type: integer, nullable: true, example: 45000 }
        materialsFils: { type: integer, nullable: true, example: 60000 }
        opsReviewedById: { type: string, format: uuid, nullable: true }
        opsReviewedAt: { type: string, format: date-time, nullable: true }
        lines: { type: array, items: { $ref: '#/components/schemas/QuoteLine' } }
        addressLine: { type: string, nullable: true }
        addressLat: { type: number, nullable: true }
        addressLng: { type: number, nullable: true }
        bookingId: { type: string, format: uuid, nullable: true }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    QuoteLine:
      type: object
      description: One itemized labour/material/prep/fee line on a quote-first BookingQuote (§17.5.3).
      properties:
        id: { type: string, format: uuid }
        quoteId: { type: string, format: uuid }
        kind: { type: string, enum: [LABOUR, MATERIAL, PREP, FEE] }
        materialId: { type: string, format: uuid, nullable: true, description: References MaterialCatalog; null = off-catalogue (routes to ops review). }
        description: { type: string }
        qty: { type: string, example: '2.00' }
        unit: { type: string, nullable: true, example: bucket }
        unitPriceFils: { type: integer, example: 12000 }
        totalFils: { type: integer, example: 24000 }
        source: { type: string, enum: [TECHNICIAN_PROCURED, CUSTOMER_SUPPLIED, PLATFORM_ARRANGED] }
        createdAt: { type: string, format: date-time }
...
    MaterialVerificationRequest:
      type: object
      description: Price-variance dispute (§17.5.14) — technician has 24h to upload the original invoice before auto-settle deducts the difference.
      properties:
        id: { type: string, format: uuid }
        bookingMaterialId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        technicianId: { type: string, format: uuid }
        status: { type: string, enum: [OPEN, INVOICE_PROVIDED, UPHELD, DEDUCTED, WITHDRAWN] }
        referencePriceFils: { type: integer }
        chargedPriceFils: { type: integer }
        deltaFils: { type: integer }
        invoiceUrl: { type: string, nullable: true }
        deadlineAt: { type: string, format: date-time }
    ConductReport:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [OFF_PLATFORM_SOLICIT, NO_SHOW, QUALITY, SAFETY, OTHER] }
        status: { type: string, enum: [OPEN, REVIEWING, UPHELD, DISMISSED] }
        subjectTechId: { type: string, format: uuid, nullable: true }
        bookingId: { type: string, format: uuid, nullable: true }
        details: { type: string, nullable: true }
        resolvedById: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a booking (customer)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingInput' }
      responses:
        '201':
          description: >
            Booking created. `checkout` is a hosted-checkout session when a real PSP is
            configured (the booking starts AWAITING_PAYMENT until the customer authorizes
            on the widget); it is null in mock/instant mode (booking is live immediately).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      booking: { $ref: '#/components/schemas/Booking' }
                      checkout:
                        oneOf:
                          - $ref: '#/components/schemas/CheckoutSession'
                          - type: 'null'
        '400': { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/schemas/Error' }
    get:
      tags: [Bookings]
      summary: List the caller's bookings
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: List of bookings.
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Booking' } } } }
...
  /admin/technicians:
    get:
      tags: [Admin]
      summary: List technicians (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: verified
          schema: { type: boolean }
        - in: query
          name: search
          description: Matches phone (contains) or name (case-insensitive contains).
          schema: { type: string }
      responses:
        '200':
          description: Paginated technicians.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Technician' } }
                  - $ref: '#/components/schemas/Meta'
...
  /admin/bookings:
    get:
      tags: [Admin]
      summary: List bookings (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Paginated bookings.
...
  /addresses:
    get:
      tags: [Addresses]
      summary: List my saved addresses
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Addresses. } }
    post:
      tags: [Addresses]
      summary: Add an address
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [label, line, lat, lng], properties: { label: { type: string }, line: { type: string }, lat: { type: number }, lng: { type: number }, building: { type: string }, apartment: { type: string }, notes: { type: string }, isDefault: { type: boolean } } } } }
      responses: { '201': { description: Created. } }
...
  /payment-methods:
    get:
      tags: [PaymentMethods]
      summary: List my saved cards (no PAN stored)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Cards. } }
    post:
      tags: [PaymentMethods]
      summary: Add a card (mock-tokenised)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [brand, last4, expMonth, expYear], properties: { brand: { type: string, enum: [visa, mastercard, amex] }, last4: { type: string }, expMonth: { type: integer }, expYear: { type: integer }, isDefault: { type: boolean } } } } }
      responses: { '201': { description: Created. } }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: My notification inbox (+ unread count in meta)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Notifications. } }
...
  /devices:
    post:
      tags: [Devices]
      summary: Register a push device token
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [token, platform], properties: { token: { type: string }, platform: { type: string, enum: [ios, android, web] } } } } }
      responses: { '201': { description: Registered. } }
...
  /guarantee:
    get:
      tags: [Guarantee]
      summary: My guarantee tickets
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Tickets. } }
    post:
      tags: [Guarantee]
      summary: Open a guarantee ticket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [bookingId, description], properties: { bookingId: { type: string, format: uuid }, description: { type: string }, mediaUrls: { type: array, items: { type: string } } } } } }
      responses: { '201': { description: Created. } }
...
  /support:
    get:
      tags: [Support]
      summary: My support tickets
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Tickets. } }
    post:
      tags: [Support]
      summary: Open a support ticket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [subject, body], properties: { subject: { type: string }, body: { type: string } } } } }
      responses: { '201': { description: Created. } }
...
  /subscriptions:
    post:
      tags: [Subscriptions]
      summary: Start the Protection plan (5 JOD/mo — priority dispatch, 15% off, 90-day guarantee, quarterly inspection)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentToken: { type: string, maxLength: 200, nullable: true, description: PSP card-on-file token for renewals; the first period is billed inline if supplied. }
      responses:
        '201':
          description: Subscription created (status ACTIVE).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Subscription' } } }
        '409': { description: Customer already has an ACTIVE subscription. }
...
  /quotes:
    post:
      tags: [Quotes]
      summary: Request a firm quote — video pre-check (fixed_scope) or quote-first assessment (mandatory for quote_first services)
      description: >-
        Exactly one of videoUrl (fixed_scope) or siteMediaUrls (quote_first) is
        required, decided by the target service's pricingModel — enforced
        server-side regardless of which is sent.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceId]
              properties:
                serviceId: { type: string, format: uuid }
                videoUrl: { type: string, format: uri, maxLength: 500, description: HTTPS URL only; required for a fixed_scope service. }
                siteMediaUrls: { type: array, maxItems: 20, items: { type: string, format: uri }, description: HTTPS URLs; required for a quote_first service. }
                dimensionsNote: { type: string, maxLength: 1000 }
                requestedTier: { type: string, enum: [ECONOMY, STANDARD, PREMIUM] }
                description: { type: string, maxLength: 1000 }
                addressLine: { type: string, maxLength: 500 }
                addressLat: { type: number, minimum: -90, maximum: 90 }
                addressLng: { type: number, minimum: -180, maximum: 180 }
      responses:
        '201':
          description: Quote created (status PENDING).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/BookingQuote' } } }
        '404': { description: Service not found. }
        '409': { description: Service is not active. }
        '422': { description: Wrong intake shape for this service's pricing archetype. }
    get:
      tags: [Quotes]
      summary: My video pre-check quotes
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Quotes.
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/BookingQuote' } } } }
...
  /verifications:
    get:
      tags: [Verifications]
      summary: My open/resolved verification requests (technician)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Requests., content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/MaterialVerificationRequest' } } } } } } } }
...
  /conduct-reports:
    post:
      tags: [Conduct]
      summary: Report off-platform solicitation, no-show, quality, or safety issues
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: { type: string, enum: [OFF_PLATFORM_SOLICIT, NO_SHOW, QUALITY, SAFETY, OTHER] }
                subjectTechId: { type: string, format: uuid }
                bookingId: { type: string, format: uuid }
                details: { type: string, maxLength: 2000 }
      responses:
        '201':
          description: Report created (status OPEN).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/ConductReport' } } }
        '404': { description: subjectTechId does not exist. }

  # ---- Referrals ----
...
  /admin/conduct-reports:
    get:
      tags: [Admin]
      summary: Conduct-report review queue (oldest first)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [OPEN, REVIEWING, UPHELD, DISMISSED] } }
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0, minimum: 0 } }
      responses:
        '200':
          description: Paginated reports.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ConductReport' } }
                  - $ref: '#/components/schemas/Meta'
...
  /admin/subscriptions:
    get:
      tags: [Admin]
      summary: Subscription list + active/past-due counts and MRR
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0, minimum: 0 } }
      responses: { '200': { description: 'Paginated subscriptions; meta adds activeCount, pastDueCount, mrrJod.' } }

  # ---- Admin: video pre-check quote pricing ----
  /admin/quotes:
    get:
      tags: [Admin]
      summary: Video pre-check quote queue
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [PENDING, QUOTED, ACCEPTED, DECLINED, EXPIRED] } }
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0, minimum: 0 } }
      responses: { '200': { description: Paginated quotes. } }
...
  /admin/verifications:
    get:
      tags: [Admin]
      summary: Price-verification disputes queue, with 24h deadline countdowns (OPS)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: query, name: status, schema: { type: string, enum: [OPEN, INVOICE_PROVIDED, UPHELD, DEDUCTED, WITHDRAWN] } }]
      responses: { '200': { description: Paginated requests., content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/MaterialVerificationRequest' } } } } } } } }
...
  /admin/guarantee:
    get:
      tags: [Admin]
      summary: Guarantee ticket queue (OPS)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: query, name: status, schema: { type: string } }]
      responses: { '200': { description: Tickets. } }
...
  /admin/support:
    get:
      tags: [Admin]
      summary: Support inbox (SUPPORT)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: query, name: status, schema: { type: string } }]
      responses: { '200': { description: Tickets. } }
...